+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Decode chunked request bodies in C
+
 ## [0.16.1] - 2026-08-23
 
 - Drop the stale `Ractors=NUM` option from the Rackup handler